CXP-1104: move connection settings queries
