+++ src/Akeneo/Connectivity/Connection/back/Infrastructure/Symfony/Resources/config/event_subscribers.yml
@@ -3,7 +3,7 @@ services:
         arguments:
             - '@akeneo_connectivity.connection.service.connection_context'
             - '@Akeneo\Connectivity\Connection\Application\Audit\Command\UpdateDataDestinationProductEventCountHandler'
-            - '@akeneo_connectivity.connection.persistence.repository.connection'
+            - '@Akeneo\Connectivity\Connection\Infrastructure\Settings\Persistence\DbalConnectionRepository'
         tags:
             - { name: kernel.event_subscriber }
