+++ src/Akeneo/Connectivity/Connection/back/Infrastructure/Symfony/Resources/config/handlers.yml
@@ -4,7 +4,7 @@ services:
         public: true
         arguments:
             - '@validator'
-            - '@akeneo_connectivity.connection.persistence.repository.connection'
+            - '@Akeneo\Connectivity\Connection\Infrastructure\Settings\Persistence\DbalConnectionRepository'
             - '@akeneo_connectivity.connection.service.client.create_client'
             - '@akeneo_connectivity.connection.service.user.create_user'
             - '@akeneo_connectivity.connection.application.handler.find_a_connection'
@@ -12,35 +12,35 @@ services:
     akeneo_connectivity.connection.application.handler.fetch_connections:
         class: 'Akeneo\Connectivity\Connection\Application\Settings\Query\FetchConnectionsHandler'
         arguments:
-            - '@akeneo_connectivity.connection.persistence.query.select_connections'
+            - '@Akeneo\Connectivity\Connection\Infrastructure\Settings\Persistence\DbalSelectConnectionsQuery'
 
     akeneo_connectivity.connection.application.handler.find_a_connection:
         class: 'Akeneo\Connectivity\Connection\Application\Settings\Query\FindAConnectionHandler'
         arguments:
-            - '@akeneo_connectivity.connection.persistence.query.select_connection_with_credentials_by_code'
+            - '@Akeneo\Connectivity\Connection\Infrastructure\Settings\Persistence\DbalSelectConnectionWithCredentialsByCodeQuery'
 
     akeneo_connectivity.connection.application.handler.update_connection:
         class: 'Akeneo\Connectivity\Connection\Application\Settings\Command\UpdateConnectionHandler'
         arguments:
             - '@validator'
-            - '@akeneo_connectivity.connection.persistence.repository.connection'
+            - '@Akeneo\Connectivity\Connection\Infrastructure\Settings\Persistence\DbalConnectionRepository'
             - '@akeneo_connectivity.connection.service.user.update_user_permissions'
 
     akeneo_connectivity.connection.application.handler.delete_connection:
         class: 'Akeneo\Connectivity\Connection\Application\Settings\Command\DeleteConnectionHandler'
         arguments:
-            - '@akeneo_connectivity.connection.persistence.repository.connection'
+            - '@Akeneo\Connectivity\Connection\Infrastructure\Settings\Persistence\DbalConnectionRepository'
             - '@akeneo_connectivity.connection.service.client.delete_client'
             - '@akeneo_connectivity.connection.service.user.delete_user'
 
     akeneo_connectivity.connection.application.handler.regenerate_connection_secret:
         class: 'Akeneo\Connectivity\Connection\Application\Settings\Command\RegenerateConnectionSecretHandler'
         arguments:
-            - '@akeneo_connectivity.connection.persistence.repository.connection'
+            - '@Akeneo\Connectivity\Connection\Infrastructure\Settings\Persistence\DbalConnectionRepository'
             - '@akeneo_connectivity.connection.service.client.regenerate_client_secret'
 
     akeneo_connectivity.connection.application.handler.regenerate_connection_password:
         class: 'Akeneo\Connectivity\Connection\Application\Settings\Command\RegenerateConnectionPasswordHandler'
         arguments:
-            - '@akeneo_connectivity.connection.persistence.repository.connection'
+            - '@Akeneo\Connectivity\Connection\Infrastructure\Settings\Persistence\DbalConnectionRepository'
             - '@akeneo_connectivity.connection.service.user.regenerate_user_password'
