--- src/Akeneo/Connectivity/Connection/back/Infrastructure/Symfony/Resources/config/queries.yml
@@ -1,15 +1,0 @@
-services:
-    akeneo_connectivity.connection.persistence.query.select_connections:
-        class: 'Akeneo\Connectivity\Connection\Infrastructure\Persistence\Dbal\Query\DbalSelectConnectionsQuery'
-        arguments:
-            - '@database_connection'
-
-    akeneo_connectivity.connection.persistence.query.select_connection_with_credentials_by_code:
-        class: 'Akeneo\Connectivity\Connection\Infrastructure\Persistence\Dbal\Query\DbalSelectConnectionWithCredentialsByCodeQuery'
-        arguments:
-            - '@database_connection'
-
-    Akeneo\Connectivity\Connection\Infrastructure\Persistence\Dbal\Query\IsConnectionsNumberLimitReachedQuery:
-        arguments:
-            - '@database_connection'
-            - '@Akeneo\Connectivity\Connection\Infrastructure\Connections\GetConnectionsNumberLimit'
