+++ src/Akeneo/Connectivity/Connection/back/Infrastructure/Symfony/Resources/config/services.yml
@@ -63,7 +63,7 @@ services:
         arguments:
             - '@akeneo_connectivity.connection.persistence.query.are_credentials_valid_combination'
             - '@akeneo_connectivity.connection.persistence.query.select_connection_code_by_client_id'
-            - '@akeneo_connectivity.connection.persistence.repository.connection'
+            - '@Akeneo\Connectivity\Connection\Infrastructure\Settings\Persistence\DbalConnectionRepository'
 
     akeneo_connectivity.connection.clock:
         class: Akeneo\Connectivity\Connection\Infrastructure\Service\Clock\SystemClock
